Get the Banff data dictionary

This function gets the data dictionary used to control the consistency of the input dataset.

get_banff_dictionary(
  version = NULL,
  which = NULL,
  language = "label:en",
  detail = FALSE
)

Arguments

version

A character string referring the version of Banff classification. The most recent classification is the default. Options are "2022" (default), "2017".

which

Indicates which variables to get from the Banff data dictionary. If NULL both input and output variables are provided.*

language

Optional input allowing the user to get the diagnoses in a specific language. Options are "label:en" (default), "label:fr", "label:de", "label:sp", "label:nl", "label:jp", "label:in".

detail

Optional argument indicating whether the data dictionary should include temporary variables generated in the process or not. FALSE by default.

Value

A list of tibbles representing meta data used in the process. The metadata are the list of variables used, and their associated categories, if any.

Examples

{

  get_banff_dictionary()

}
#> $ReadMe
#> # A tibble: 15 × 4
#>    `Data Dictionary`             ...2       ...3                           ...4 
#>    <chr>                         <chr>      <chr>                          <chr>
#>  1 Authors:                      NA         R Sapir-Pichhadze, E Gitelman… NA   
#>  2 NA                            NA         NA                             NA   
#>  3 Aim of the document:          NA         Provide a standard metadata m… NA   
#>  4 NA                            NA         NA                             NA   
#>  5 Current version:              NA         Version 2.0.0                  NA   
#>  6 Banff Classification version: NA         2022                           NA   
#>  7 NA                            NA         NA                             NA   
#>  8 Spreadsheets:                 NA         Description                    NA   
#>  9 NA                            ReadMe     Description of the document a… NA   
#> 10 NA                            Variables  Metadata of the list of minim… NA   
#> 11 NA                            Categories Code, label, and description … NA   
#> 12 NA                            NA         NA                             NA   
#> 13 Version history:              NA         Description                    Date 
#> 14 NA                            1.0.0      Initial version of the data d… 45415
#> 15 NA                            2.0.0      Version of the data dictionna… 45519
#> 
#> $Variables
#> # A tibble: 22 × 9
#>    index name      label valueType categorical input output `NA accepted or not`
#>    <dbl> <chr>     <chr> <chr>     <chr>       <chr> <chr>  <chr>               
#>  1     1 patient_… "Pat… text      no          yes   yes    NO                  
#>  2     2 center    "Cen… text      no          yes   yes    NO                  
#>  3     3 biopsy_id "Bio… text      no          yes   yes    NO                  
#>  4     4 sc_date_… "Dat… date      no          yes   yes    NO                  
#>  5     5 date_tx   "Tra… date      no          yes   yes    YES                 
#>  6     6 type_bx   "Typ… integer   yes         yes   yes    YES                 
#>  7   105 diag_cod… "Cat… integer   yes         no    yes    Not Applicable      
#>  8   106 diag_cod… "Cat… integer   yes         no    yes    Not Applicable      
#>  9   107 diag_cod… "Cat… integer   yes         no    yes    Not Applicable      
#> 10   108 diag_cod… "Cat… integer   yes         no    yes    Not Applicable      
#> # ℹ 12 more rows
#> # ℹ 1 more variable: constraints <chr>
#> 
#> $Categories
#> # A tibble: 44 × 4
#>    variable    name  labels                                          description
#>    <chr>       <chr> <chr>                                           <chr>      
#>  1 type_bx     1     Surveillance Biopsy                             Surveillan…
#>  2 type_bx     2     Biopsy for Cause                                Biopsy for…
#>  3 diag_code_1 1     Category 1: Normal Biopsy or Nonspecific Chang… Category 1…
#>  4 diag_code_2 1     Category 2.1: C4d Staining Without Evidence of… same       
#>  5 diag_code_2 2     Category 2.2: C4d Staining With Acute Tubular … new        
#>  6 diag_code_2 3     Category 2.3: Accommodation                     new        
#>  7 diag_code_2 4     Category 2.4: Probable AMR                      new        
#>  8 diag_code_2 5     Category 2.5: Microvascular Inflammation/Injur… new        
#>  9 diag_code_2 6     Category 2.6: Suspicious for Active AMR         was 2      
#> 10 diag_code_2 7     Category 2.7: Active AMR                        was 3      
#> # ℹ 34 more rows
#>